+++ .github/workflows/AutoPushToPypi.yml
@@ -2,7 +2,6 @@ name: Push rapid_videocr to pypi
 
 on:
   push:
-    # branches: [ main ]
     tags:
       - v*
