+++ README.md
@@ -17,22 +17,25 @@
 </div>
 
 ### Introduction
+
 - Video hard subtitle extraction, automatically generate the corresponding `srt | txt` file.
 - Supported subtitle languages: Chinese | English (For other supported languages, see: [List of supported languages](https://github.com/PaddlePaddle/PaddleOCR/blob/release/2.1/doc/doc_ch/multi_languages.md#%E8%AF%AD%E7%A7%8D%E7%BC%A9%E5%86%99))
 - The advantages are as follows:
-   - **Faster extraction**: Used in conjunction with [VideoSubFinder](https://sourceforge.net/projects/videosubfinder/) software to extract key subtitle frames faster.
-   - **More accurate recognition**: Use [RapidOCR](https://github.com/RapidAI/RapidOCR) as the recognition library.
-   - **More convenient to use**: pip can be installed directly and used.
+    - **Faster extraction**: Used in conjunction with [VideoSubFinder](https://sourceforge.net/projects/videosubfinder/) software to extract key subtitle frames faster.
+    - **More accurate recognition**: Use [RapidOCR](https://github.com/RapidAI/RapidOCR) as the recognition library.
+    - **More convenient to use**: pip can be installed directly and used.
 
 - For desktop EXE version, please go to [RapidVideOCRDesktop](https://github.com/SWHL/RapidVideOCRDesktop).
 - If it helps you, please give a star ⭐.
 
 ### [Online Demo](https://huggingface.co/spaces/SWHL/RapidVideOCR)
+
 <div align="center">
     <img src="https://github.com/SWHL/RapidVideOCR/releases/download/v2.0.1/OnlineDemo.gif" alt="Demo" width="100%" height="100%">
 </div>
 
 ### Overall framework
+
 ```mermaid
 flowchart LR
     A[/Video/] --Extract subtitle key frame--> B(VideoSubFinder) --OCR-->C(RapidVideOCR)
@@ -40,6 +43,7 @@ flowchart LR
 ```
 
 ### Installation
+
 ```bash
 pip install rapid_videocr
 ```
@@ -55,9 +59,11 @@ rapid_videocr -i RGBImages
 ```
 
 ### Documentation
+
 Full documentation can be found on [docs](https://swhl.github.io/RapidVideOCR/docs), in Chinese.
 
 ### Code Contributors
+
 <p align="left">
   <a href="https://github.com/SWHL/RapidVideOCR/graphs/contributors">
     <img src="https://contrib.rocks/image?repo=SWHL/RapidVideOCR" width="20%"/>
@@ -65,6 +71,7 @@ Full documentation can be found on [docs](https://swhl.github.io/RapidVideOCR/do
 </p>
 
 ### Contributing
+
 - Pull requests are welcome. For major changes, please open an issue first
 to discuss what you would like to change.
 - Please make sure to update tests as appropriate.
@@ -78,4 +85,5 @@ If you want to sponsor the project, you can directly click the **Buy me a coffee
 </div>
 
 ### License
+
 This project is released under the [Apache 2.0 license](./LICENSE).
