+++ rapid_videocr/__init__.py
@@ -1,6 +1,5 @@
 # -*- encoding: utf-8 -*-
 # @Author: SWHL
 # @Contact: liekkaskono@163.com
-from .logger import logger
-from .main import RapidVideoSubFinderOCR
-from .rapid_videocr import RapidVideOCR, RapidVideOCRError
+from .vsf_ocr_cli import RapidVideoSubFinderOCR
+from .main import RapidVideOCR, RapidVideOCRExeception
