+++ rapid_videocr/main.py
@@ -2,144 +2,358 @@
 # @Author: SWHL
 # @Contact: liekkaskono@163.com
 import argparse
+from enum import Enum
 from pathlib import Path
-from typing import Optional
+from typing import Any, Dict, List, Optional, Tuple, Union
+
+import cv2
+import numpy as np
+from rapidocr import RapidOCR
+from tqdm import tqdm
 
 try:
-    from .logger import logger
-    from .rapid_videocr import RapidVideOCR
-    from .utils import float_range
-    from .video_sub_finder import VideoSubFinder
+    from .utils.logger import Logger
+    from .utils.utils import (
+        CropByProject,
+        compute_centroid,
+        compute_poly_iou,
+        is_inclusive_each_other,
+        mkdir,
+        padding_img,
+        read_img,
+        write_txt,
+    )
 except:
-    from logger import logger
-    from utils import float_range
-    from video_sub_finder import VideoSubFinder
+    from utils.utils import (
+        CropByProject,
+        compute_centroid,
+        compute_poly_iou,
+        is_inclusive_each_other,
+        mkdir,
+        padding_img,
+        read_img,
+        write_txt,
+    )
 
-    from rapid_videocr import RapidVideOCR
+    from rapid_videocr.utils.logger import Logger
+
+CUR_DIR = Path(__file__).resolve().parent
 
 
-class RapidVideoSubFinderOCR:
+class OutputFormat(Enum):
+    TXT = "txt"
+    SRT = "srt"
+    ALL = "all"
+
+
+class RapidVideOCR:
     def __init__(
         self,
         is_concat_rec: bool = False,
         concat_batch: int = 10,
-        out_format: str = "all",
+        out_format: str = OutputFormat.ALL.value,
         is_print_console: bool = False,
-        vsf_exe_path: Optional[str] = None,
-        clear_dirs: bool = True,
-        run_search: bool = True,
-        create_cleared_text_images: bool = True,
-        create_empty_sub: Optional[str] = None,
-        create_sub_from_cleared_txt_images: Optional[str] = None,
-        create_sub_from_txt_results: Optional[str] = None,
-        open_video_opencv: bool = True,
-        open_video_ffmpeg: bool = False,
-        use_cuda: bool = False,
-        start_time: Optional[str] = None,
-        end_time: Optional[str] = None,
-        top_video_image_percent_end: float = 0.2,
-        bottom_video_image_percent_end: float = 0.0,
-        left_video_image_percent_end: float = 0.0,
-        right_video_image_percent_end: float = 1.0,
-        general_settings: Optional[str] = None,
-        num_threads: int = -1,
-        num_ocr_threads: int = 1,
-        **kwargs,
-    ) -> None:
-        self.vsf = VideoSubFinder(
-            vsf_exe_path,
-            clear_dirs,
-            run_search,
-            create_cleared_text_images,
-            create_empty_sub,
-            create_sub_from_cleared_txt_images,
-            create_sub_from_txt_results,
-            open_video_opencv,
-            open_video_ffmpeg,
-            use_cuda,
-            start_time,
-            end_time,
-            top_video_image_percent_end,
-            bottom_video_image_percent_end,
-            left_video_image_percent_end,
-            right_video_image_percent_end,
-            general_settings,
-            num_threads,
-            num_ocr_threads,
-        )
+        ocr_params: Optional[Dict[str, Any]] = None,
+    ):
+        self.logger = Logger(logger_name=__name__).get_log()
+        self.ocr_processor = OCRProcessor(ocr_params, concat_batch)
+        self.cropper = CropByProject()
+        self.is_concat_rec = is_concat_rec
+        self.out_format = out_format
+        self.is_print_console = is_print_console
 
-        self.video_ocr = RapidVideOCR(
-            is_concat_rec=is_concat_rec,
-            concat_batch=concat_batch,
-            out_format=out_format,
-            is_print_console=is_print_console,
-            **kwargs,
-        )
-        self.video_formats = [".mp4", ".avi", ".mov", ".mkv"]
+    def __call__(
+        self,
+        video_sub_finder_dir: Union[str, Path],
+        save_dir: Union[str, Path],
+        save_name: str = "result",
+    ) -> List[str]:
+        save_dir = Path(save_dir)
+        mkdir(save_dir)
 
-    def __call__(self, video_path: str, output_dir: str = "outputs"):
-        if Path(video_path).is_dir():
-            video_list = Path(video_path).rglob("*.*")
-            video_list = [
-                v for v in video_list if v.suffix.lower() in self.video_formats
-            ]
-        else:
-            video_list = [video_path]
-
-        logger.info(
-            "Extracting subtitle images with VideoSubFinder (takes quite a long time) ..."
-        )
-        video_num = len(video_list)
-        for i, one_video in enumerate(video_list):
-            logger.info(
-                "[%s/%s] Starting to extract %s key frame", i + 1, video_num, one_video
+        img_list = self.get_img_list(video_sub_finder_dir)
+        if len(img_list) <= 0:
+            raise RapidVideOCRExeception(
+                f"{video_sub_finder_dir} does not have valid images"
             )
 
-            save_name = Path(one_video).stem
-            save_dir = Path(output_dir) / save_name
-            save_vsf_dir = save_dir / "VSF_Results"
+        srt_result, txt_result = self.ocr_processor(
+            img_list, self.is_concat_rec, self.is_txt_dir
+        )
+        self.export_file(save_dir, save_name, srt_result, txt_result)
+        if self.is_print_console:
+            self.print_console(txt_result)
 
-            try:
-                self.vsf(str(one_video), str(save_vsf_dir))
-            except Exception as e:
-                logger.error("Extract %s error, %s, skip", one_video, e)
+        return txt_result
+
+    def get_img_list(self, video_sub_finder_dir: Union[str, Path]) -> List[Path]:
+        def get_sort_key(x: Path) -> int:
+            return int("".join(str(x.stem).split("_")[:4]))
+
+        video_sub_finder_dir = Path(video_sub_finder_dir)
+        if not video_sub_finder_dir.exists():
+            raise RapidVideOCRExeception(f"{video_sub_finder_dir} does not exist.")
+
+        self.is_txt_dir = "TXTImages" in Path(video_sub_finder_dir).name
+
+        img_list = list(Path(video_sub_finder_dir).glob("*.jpeg"))
+        img_list = sorted(img_list, key=get_sort_key)
+        if not img_list:
+            return []
+        return img_list
+
+    def export_file(
+        self,
+        save_dir: Path,
+        save_name: str,
+        srt_result: List[str],
+        txt_result: List[str],
+    ):
+        format_actions = {
+            OutputFormat.TXT.value: [(OutputFormat.TXT.value, txt_result)],
+            OutputFormat.SRT.value: [(OutputFormat.SRT.value, srt_result)],
+            OutputFormat.ALL.value: [
+                (OutputFormat.TXT.value, txt_result),
+                (OutputFormat.SRT.value, srt_result),
+            ],
+        }
+
+        try:
+            actions = format_actions[self.out_format]
+        except KeyError:
+            raise ValueError(f"Unsupported output format: {self.out_format}") from None
+
+        for ext, content in actions:
+            file_path = save_dir / f"{save_name}.{ext}"
+            write_txt(file_path, content)
+
+        self.logger.info("[OCR] Results saved to directory: %s", save_dir)
+
+    def print_console(self, txt_result: List):
+        for v in txt_result:
+            print(v.strip())
+
+
+class OCRProcessor:
+    def __init__(self, ocr_params, batch_size: int):
+        self.logger = Logger(logger_name=__name__).get_log()
+
+        default_params = {"Global.width_height_ratio": -1}
+        if ocr_params is not None:
+            ocr_params.update(default_params)
+        else:
+            ocr_params = default_params
+        self.ocr_engine = RapidOCR(params=ocr_params)
+
+        self.batch_size = batch_size
+
+    def __call__(self, img_list: List[Path], is_concat_rec: bool, is_txt_dir: bool):
+        self.is_txt_dir = is_txt_dir
+        process_func = self.batch_rec if is_concat_rec else self.single_rec
+        srt_result, txt_result = process_func(img_list)
+        return srt_result, txt_result
+
+    def single_rec(self, img_list: List[Path]) -> Tuple[List[str], List[str]]:
+        self.logger.info("[OCR] Running with single recognition.")
+
+        srt_result, txt_result = [], []
+        for i, img_path in enumerate(tqdm(img_list, desc="OCR")):
+            time_str = self.get_srt_timestamp(img_path)
+
+            img = read_img(img_path)
+            img = padding_img(img, (img.shape[0], img.shape[0], 0, 0))
+
+            dt_boxes, rec_res = self.get_ocr_result(img)
+            if dt_boxes is None or rec_res is None:
+                txts = ""
+            else:
+                txts = self.process_same_line(dt_boxes, rec_res)
+
+            srt_result.append(f"{i+1}\n{time_str}\n{txts}\n")
+            txt_result.append(f"{txts}\n")
+        return srt_result, txt_result
+
+    @staticmethod
+    def get_srt_timestamp(file_path: Path) -> str:
+        split_paths = file_path.stem.split("_")
+        start_time = split_paths[:4]
+        start_time[0] = f"{start_time[0]:0>2}"
+        start_str = ":".join(start_time[:3]) + f",{start_time[3]}"
+
+        end_time = split_paths[5:9]
+        end_time[0] = f"{end_time[0]:0>2}"
+        end_str = ":".join(end_time[:3]) + f",{end_time[3]}"
+        return f"{start_str} --> {end_str}"
+
+    def batch_rec(self, img_list: List[Path]) -> Tuple[List, List]:
+        self.logger.info("[OCR] Running with concat recognition.")
+
+        img_nums = len(img_list)
+        srt_result, txt_result = [], []
+        for start_i in tqdm(range(0, img_nums, self.batch_size), desc="Concat Rec"):
+            end_i = min(img_nums, start_i + self.batch_size)
+
+            concat_img, img_coordinates, img_paths = self.get_batch_img_info(
+                img_list, start_i, end_i
+            )
+            dt_boxes, rec_res = self.get_ocr_result(concat_img)
+            if rec_res is None or dt_boxes is None:
                 continue
 
-            logger.info("[%s/%s] Starting to run %s ocr", i + 1, video_num, one_video)
+            srt_part, txt_part = self.get_match_results(
+                start_i, img_coordinates, dt_boxes, rec_res, img_paths
+            )
+            srt_result.extend(srt_part)
+            txt_result.extend(txt_part)
+        return srt_result, txt_result
 
-            rgb_dir = Path(save_vsf_dir) / "RGBImages"
-            if not list(rgb_dir.iterdir()):
-                logger.warning("Extracting frames from %s is 0, skip", one_video)
-                continue
-            self.video_ocr(rgb_dir, save_dir, save_name=save_name)
+    def get_batch_img_info(
+        self, img_list: List[Path], start: int, end: int
+    ) -> Tuple[np.ndarray, np.ndarray, List[Path]]:
+        select_imgs = img_list[start:end]
+
+        padding_value = 10
+        array_img_list, img_coordinates, batch_img_paths = [], [], []
+        for i, img_path in enumerate(select_imgs):
+            batch_img_paths.append(img_path)
+
+            img = read_img(img_path)
+
+            if self.is_txt_dir:
+                img = cv2.resize(img, None, fx=0.25, fy=0.25)
+
+            pad_img = padding_img(img, (0, padding_value, 0, 0))
+            array_img_list.append(pad_img)
+
+            h, w = img.shape[:2]
+            x0, y0 = 0, i * (h + padding_value)
+            x1, y1 = w, (i + 1) * (h + padding_value)
+            img_coordinates.append([(x0, y0), (x1, y0), (x1, y1), (x0, y1)])
+
+        concat_img = np.vstack(array_img_list)
+        return concat_img, np.array(img_coordinates), batch_img_paths
+
+    def get_match_results(
+        self,
+        start_i: int,
+        img_coordinates: np.ndarray,
+        dt_boxes: np.ndarray,
+        rec_res: Tuple[str],
+        img_paths: List[Path],
+    ) -> Tuple[List[str], List[str]]:
+        match_dict: Dict[int, List[Union[Path, np.ndarray, str]]] = {
+            k: [] for k in range(len(img_coordinates))
+        }
+        visited_idx = []
+        for i, frame_boxes in enumerate(img_coordinates):
+            for idx, dt_box in enumerate(dt_boxes):
+                if idx in visited_idx:
+                    continue
+
+                box_iou = compute_poly_iou(frame_boxes, dt_box)
+                if is_inclusive_each_other(frame_boxes, dt_box) or box_iou > 0.1:
+                    matched_path = img_paths[i]
+                    match_dict.setdefault(i, []).append(
+                        [matched_path, dt_box, rec_res[idx]]
+                    )
+                    visited_idx.append(idx)
+
+        srt_result_part, txt_result_part = [], []
+        for k, v in match_dict.items():
+            cur_frame_idx = start_i + k
+            if v:
+                img_path, boxes, recs = list(zip(*v))
+                time_str = self.get_srt_timestamp(img_path[0])
+                txts = self.process_same_line(boxes, recs)
+            else:
+                time_str = self.get_srt_timestamp(img_paths[k])
+                txts = ""
+
+            srt_result_part.append(f"{cur_frame_idx+1}\n{time_str}\n{txts}\n")
+            txt_result_part.append(f"{txts}\n")
+        return srt_result_part, txt_result_part
+
+    def get_ocr_result(
+        self, img: np.ndarray
+    ) -> Tuple[Optional[np.ndarray], Optional[Tuple[str]]]:
+        ocr_result = self.ocr_engine(img)
+        if ocr_result.boxes is None:
+            return None, None
+        return ocr_result.boxes, ocr_result.txts
+
+    def process_same_line(self, dt_boxes: np.ndarray, rec_res: List[str]) -> str:
+        """处理同一行文本的识别结果
+
+        Args:
+            dt_boxes: 检测框坐标数组
+            rec_res: 识别结果列表
+
+        Returns:
+            合并后的文本结果字符串
+        """
+        if len(rec_res) == 1:
+            return rec_res[0]
+
+        y_centroids = [compute_centroid(box)[1] for box in dt_boxes]
+        line_groups = self._group_by_lines(y_centroids)
+        return self._merge_line_text(line_groups, rec_res)
+
+    def _group_by_lines(self, y_centroids: List[float]) -> List[List[int]]:
+        """将文本框按行分组"""
+        bool_res = self.is_same_line(y_centroids)
+        groups = []
+        current_group = [0]
+        for i, is_same in enumerate(bool_res, 1):
+            if is_same:
+                current_group.append(i)
+            else:
+                groups.append(current_group)
+                current_group = [i]
+
+        groups.append(current_group)
+        return groups
+
+    def _merge_line_text(self, line_groups: List[List[int]], rec_res: List[str]) -> str:
+        lines = []
+        for group in line_groups:
+            line_text = " ".join(rec_res[i] for i in group)
+            lines.append(line_text)
+        return "\n".join(lines)
+
+    @staticmethod
+    def is_same_line(points: List) -> List[bool]:
+        threshold = 5
+
+        align_points = list(zip(points, points[1:]))
+        bool_res = [False] * len(align_points)
+        for i, point in enumerate(align_points):
+            y0, y1 = point
+            if abs(y0 - y1) <= threshold:
+                bool_res[i] = True
+        return bool_res
+
+
+class RapidVideOCRExeception(Exception):
+    pass
 
 
 def main():
     parser = argparse.ArgumentParser()
-
-    videocr_param_group = parser.add_argument_group(title="VideOCRParameters")
-    videocr_param_group.add_argument(
-        "-video_dir",
-        "--video_dir",
-        type=str,
-        default=None,
-        help="The full path of video or the path of video directory.",
-    )
-    videocr_param_group.add_argument(
+    parser.add_argument(
         "-i",
         "--img_dir",
         type=str,
-        default=None,
+        required=True,
         help="The full path of RGBImages or TXTImages.",
     )
-    videocr_param_group.add_argument(
+    parser.add_argument(
         "-s",
         "--save_dir",
         type=str,
         default="outputs",
         help='The path of saving the recognition result. Default is "outputs" under the current directory.',
     )
-    videocr_param_group.add_argument(
+    parser.add_argument(
         "-o",
         "--out_format",
         type=str,
@@ -147,169 +361,36 @@ def main():
         choices=["srt", "txt", "all"],
         help='Output file format. Default is "all".',
     )
-    videocr_param_group.add_argument(
+    parser.add_argument(
         "--is_concat_rec",
         action="store_true",
         default=False,
         help="Which mode to run (concat recognition or single recognition). Default is False.",
     )
-    videocr_param_group.add_argument(
+    parser.add_argument(
         "-b",
         "--concat_batch",
         type=int,
         default=10,
         help="The batch of concating image nums in concat recognition mode. Default is 10.",
     )
-    videocr_param_group.add_argument(
+    parser.add_argument(
         "-p",
-        "--is_print_console",
-        action="store_true",
-        default=False,
-        help="Whether to print the subtitle results to console. -p means to print.",
-    )
-
-    vsf_param_group = parser.add_argument_group(title="VSFParameters")
-    vsf_param_group.add_argument(
-        "-vsf",
-        "--vsf_exe_path",
-        type=str,
-        default=None,
-        help="The full path of VideoSubFinderWXW.exe.",
-    )
-    vsf_param_group.add_argument(
-        "-c",
-        "--clear_dirs",
-        action="store_false",
-        default=True,
-        help="Clear Folders (remove all images), performed before any other steps. Default is True",
-    )
-    vsf_param_group.add_argument(
-        "-r",
-        "--run_search",
-        action="store_false",
-        default=True,
-        help="Run Search (find frames with hardcoded text (hardsub) on video) Default is True",
-    )
-    vsf_param_group.add_argument(
-        "-ccti",
-        "--create_cleared_text_images",
-        action="store_true",
-        default=False,
-        help="Create Cleared Text Images. Default is True",
-    )
-    vsf_param_group.add_argument(
-        "-ces",
-        "--create_empty_sub",
-        type=str,
-        default=None,
-        help="Create Empty Sub With Provided Output File Name (*.ass or *.srt)",
-    )
-    vsf_param_group.add_argument(
-        "-cscti",
-        "--create_sub_from_cleared_txt_images",
-        type=str,
-        default=None,
-        help="Create Sub From Cleared TXT Images With Provided Output File Name (*.ass or *.srt)",
-    )
-    vsf_param_group.add_argument(
-        "-cstxt",
-        "--create_sub_from_txt_results",
-        type=str,
-        default=None,
-        help="Create Sub From TXT Results With Provided Output File Name (*.ass or *.srt)",
-    )
-    vsf_param_group.add_argument(
-        "-ovocv",
-        "--open_video_opencv",
-        action="store_false",
-        default=True,
-        help="open video by OpenCV (default). Default is True",
-    )
-    vsf_param_group.add_argument(
-        "-ovffmpeg",
-        "--open_video_ffmpeg",
-        action="store_true",
-        default=False,
-        help="open video by FFMPEG",
-    )
-    vsf_param_group.add_argument(
-        "-uc", "--use_cuda", action="store_true", default=False, help="use cuda"
-    )
-    vsf_param_group.add_argument(
-        "--start_time",
-        type=str,
-        default="0:00:00:000",
-        help="start time, default = 0:00:00:000 (in format hour:min:sec:milisec)",
-    )
-    vsf_param_group.add_argument(
-        "--end_time",
-        type=str,
-        default=None,
-        help="end time, default = video length",
-    )
-    vsf_param_group.add_argument(
-        "-te",
-        "--top_video_image_percent_end",
-        type=float_range(0, 1.0),
-        default=0.2,
-        help="top video image percent offset from image bottom, can be in range [0.0,1.0], default = 1.0",
-    )
-    vsf_param_group.add_argument(
-        "-be",
-        "--bottom_video_image_percent_end",
-        type=float_range(0, 1.0),
-        default=0.0,
-        help="bottom video image percent offset from image bottom, can be in range [0.0,1.0], default = 0.0",
-    )
-    vsf_param_group.add_argument(
-        "-le",
-        "--left_video_image_percent_end",
-        type=float_range(0, 1.0),
-        default=0.0,
-        help="left video image percent end, can be in range [0.0,1.0], default = 0.0",
-    )
-    vsf_param_group.add_argument(
-        "-re",
-        "--right_video_image_percent_end",
-        type=float_range(0, 1.0),
-        default=1.0,
-        help="right video image percent end, can be in range [0.0,1.0], default = 1.0",
-    )
-    vsf_param_group.add_argument(
-        "-gs",
-        "--general_settings",
-        default=None,
-        help="general settings (path to general settings *.cfg file, default = settings/general.cfg)",
-    )
-    vsf_param_group.add_argument(
-        "-nthr",
-        "--num_threads",
-        type=int,
-        default=1,
-        help="number of threads used for Run Search",
-    )
-    vsf_param_group.add_argument(
-        "-nocrthr",
-        "--num_ocr_threads",
-        type=int,
-        default=1,
-        help="number of threads used for Create Cleared TXT Images",
+        "--print_console",
+        type=bool,
+        default=0,
+        choices=[0, 1],
+        help="Whether to print the subtitle results to console. 1 means to print results to console. Default is 0.",
     )
     args = parser.parse_args()
 
-    if args.vsf_exe_path and args.video_dir:
-        extractor = RapidVideoSubFinderOCR(**vars(args))
-        extractor(args.video_dir, args.save_dir)
-    elif args.img_dir:
-        extractor = RapidVideOCR(
-            is_concat_rec=args.is_concat_rec,
-            concat_batch=args.concat_batch,
-            out_format=args.out_format,
-            is_print_console=args.is_print_console,
-        )
-        extractor(args.img_dir, args.save_dir)
-    else:
-        pass
+    extractor = RapidVideOCR(
+        is_concat_rec=args.is_concat_rec,
+        concat_batch=args.concat_batch,
+        out_format=args.out_format,
+        is_print_console=args.print_console,
+    )
+    extractor(args.img_dir, args.save_dir)
 
 
 if __name__ == "__main__":
